feat: add patch-documents.js, update README
